Rename rout names to fit flat folter structure standards
