+++ src/components/TimelineProposal.vue
@@ -28,7 +28,7 @@ watchEffect(() => {
   <router-link
     class="p-4 block text-color"
     :to="{
-      name: 'proposal',
+      name: 'SpaceProposal',
       params: { key: proposal.space.id, id: proposal.id }
     }"
   >
