+++ src/views/SpaceSettings.vue
@@ -223,9 +223,9 @@ onMounted(async () => {
   <Layout v-bind="$attrs">
     <template #content-left>
       <div class="px-4 md:px-0 mb-3">
-        <router-link :to="{ name: 'home' }" class="text-color">
+        <router-link :to="{ name: 'SpaceProposals' }" class="text-color">
           <Icon name="back" size="22" class="!align-middle" />
-          {{ $t('backToHome') }}
+          {{ space.name }}
         </router-link>
       </div>
       <div class="px-4 md:px-0">
